dnet: Factor out the buffering stuff from acloud_reverseconn
So that it can be reused by other net.Conn implementations.
